--- src/components/About.tsx
@@ -1,27 +1,0 @@
-export const About = () => {
-  return (
-    <section className="leading-8">
-      <h2 className="header2">About Me</h2>
-      <p className="mb-8 text-darkgray">
-        東京在住のWebエンジニア兼教師。大学で初等教育を専攻している中で、これからの情報社会の隆盛を感じ、これから子供たちが経験していくであろう世界を自ら知るためITの世界へ飛び込む。
-      </p>
-      <p className="text-darkgray">
-        2017年から医療・ヘルスケア領域の受託開発・研究開発や学校向け教育プラットフォームの自社開発・運用までWebアプリケーションを中心に幅広く携わる。
-        <br />
-        2022年現在はエンジニアを休みながら教師として活動中。
-      </p>
-      <p className="text-darkgray">
-        その他最近のエンジニアリング絡みの活動などは
-        <a
-          href="https://lapras.com/public/WBVHFUW"
-          target="_blank"
-          rel="noreferrer"
-          className="text-blue-600 hover:underline"
-        >
-          lapras
-        </a>
-        をご覧ください。
-      </p>
-    </section>
-  );
-};
