--- src/components/AboutHeader.tsx
@@ -1,11 +1,0 @@
-import { HeaderBar } from './HeaderBar';
-import { HeaderImage } from './HeaderImage';
-
-export const AboutHeader = () => {
-  return (
-    <header>
-      <HeaderBar />
-      <HeaderImage />
-    </header>
-  );
-};
