+++ src/components/BlogHeader.tsx
@@ -3,7 +3,7 @@ import { HeaderBar } from './HeaderBar';
 export const BlogHeader = () => {
   return (
     <header>
-      <HeaderBar type="blog" />
+      <HeaderBar />
     </header>
   );
 };
