+++ src/lib/site-info.ts
@@ -1,2 +1,2 @@
-export const siteTitle = "Hakshu's Portfolio";
+export const siteTitle = "Hakshu's Blog";
 export const siteUrl: string = process.env.NEXT_PUBLIC_VERCEL_URL || '';
